Original Optimum Silhouette Clustering Algorithm

OSil implements the original Optimum Silhouette algorithm.

OSil(dist, k, initClustering = NULL, initMethod = "average")

Arguments

dist

A "dist" object, which can be obtained by the "dist" function.

k

The number of clusters.

initClustering

A user-specified initialized clustering. It must be an integer vector of k unique values 1,2,...,k. By defaualt, initClustering is set to NULL. If initClustering is NULL, initMethod is used; otherwise, initClustering is used.

initMethod

An initialization method. By default, effOSil uses average-linkage clustering ("average") for initialization. Other options include well-established clustering algorithms such as Partition Around Medoids ("pam"), single-linkage clustering ("single"), complete-linkage clustering ("complete"), the Ward's method ("ward.D"), etc.

Value

Clustering

The OSil clustering solution.

ASW

The ASW associated with the OSil clustering.

nIter

The number of iterations needed for convergence.

References

Batool F. (2019). Optimum average silhouette width clustering. PhD Thesis, University College London.

Author

Minh Long Nguyen edelweiss611428@gmail.com